8325525 Addressed review comments
