+++ test/jdk/tools/jpackage/apps/ThirdPartyAppLauncher.java
@@ -21,49 +21,17 @@
  * questions.
  */
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.logging.Logger;
 
 public class ThirdPartyAppLauncher {
     private static final Logger logger = Logger
             .getLogger(ThirdPartyAppLauncher.class.getName());
-    private static final String FS = File.separator;
-    private static final Path CWD = Paths.get(".");
-    private static final String processIdFilePath = CWD + FS + "process.tmp";
 
     public static void main(String[] args) throws IOException {
-
         ProcessBuilder processBuilder = new ProcessBuilder("regedit");
         Process process = processBuilder.start();
         logger.info("RegEdit id=" + process.pid());
-
-        File file = new File(processIdFilePath);
-        BufferedWriter writer = null;
-        try {
-            if (file.createNewFile()) {
-                logger.info(file.getAbsolutePath().toString());
-                writer = new BufferedWriter(new FileWriter(file));
-                writer.write("" + process.pid());
-                writer.close();
-                logger.info("Successfully written processid to file "
-                        + processIdFilePath);
-            }
-        } catch (IOException e) {
-            throw new IOException(e);
-        } finally {
-            try {
-                if (writer != null) {
-                    writer.close();
-                }
-            } catch (IOException ex) {
-                throw new IOException(ex);
-            }
-        }
         System.exit(0);
     }
 }
